requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
black==22.6.0
blinker==1.5
Brotli==1.0.9
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.5
contourpy==1.0.6
cryptography==39.0.0
customtkinter==5.0.5
cycler==0.11.0
darkdetect==0.8.0
exceptiongroup==1.1.0
fonttools==4.38.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
kaitaistruct==0.10
kiwisolver==1.4.4
matplotlib==3.6.2
mypy-extensions==0.4.3
numpy==1.23.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
outcome==1.2.0
packaging==21.3
pandas==1.5.3
pathspec==0.10.1
Pillow==9.3.0
platformdirs==2.5.2
pyasn1==0.4.8
pycparser==2.21
pydivert==2.1.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
regex==2021.11.10
requests==2.28.2
selenium==4.8.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tk==0.1.0
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
webdriver-manager==3.8.5
wpiformat==2022.25
wsproto==1.2.0
zstandard==0.19.0
